kernel/binary_manager: Rename functions
- Rename functions and used variables for readability and better understanding.
- Move functions to proper location.
